Split grpc/wrappers.go into several files.
There is now one file per service, containing both the client-side and
server-side wrappers for that service. This is a straight move of the code, with
the copyright, header comments, package statement, and imports copied into each
new file, and `goimports` run on the result.

Two custom errors were moved into bcodes.go.
